feat(#973): Refactor AvailabilityCalendar into composable components
Split monolithic AvailabilityCalendar into:
- useAvailabilityCalendar hook for state management and logic
- CalendarHeader component for month navigation
- CalendarGrid component for calendar grid display
- CalendarLegend component for legend display
- SlotEditor component for managing availability slots
- SelectionSummary component for date range display

Improves maintainability, testability, and code reusability.
